python-ai-service/requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
asyncpg==0.29.0

# AI and ML
openai==1.12.0
anthropic==0.18.1
transformers==4.37.2
sentence-transformers==2.3.1
langchain==0.1.5
langchain-anthropic==0.1.1
tiktoken==0.5.2

# Vector Database
pinecone-client==4.1.2
weaviate-client==4.4.3

# Embeddings and NLP
numpy==1.26.4
scipy==1.11.4  # Compatible with Python 3.12
scikit-learn==1.3.2  # Compatible with Python 3.12
spacy==3.7.2
torch==2.1.2  # Required for sentence-transformers
torchvision==0.16.2  # May be needed for some models

# API and HTTP
httpx==0.26.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Environment and Config
python-dotenv==1.0.0

# Logging and Monitoring
loguru==0.7.2
prometheus-client==0.19.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0

# CORS and Security
slowapi==0.1.9

# Date and Time
python-dateutil==2.8.2

# Redis for caching
redis==5.0.1
aioredis==2.0.1

# Background Tasks
celery==5.3.5
flower==2.0.1

# JSON handling
orjson==3.9.10

# Graph and Network Analysis
networkx==3.2.1
python-louvain==0.16  # For community detection

# Advanced Memory System Dependencies
# pgvector==0.2.4  # Uncomment when PostgreSQL vector extension is available

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0